Rename to ChangeNotificationOwner service
